[RFR] Improve patchwork
