+++ upgrade_analysis/__init__.py
@@ -4,5 +4,4 @@ from . import wizards
 from . import blacklist
 from . import apriori
 from . import compare
-from . import upgrade_loading
 from . import upgrade_log
