+++ upgrade_analysis/odoo_patch/__init__.py
@@ -1,2 +1,3 @@
-from . import odoo
 from . import addons
+from . import odoo
+from . import odoo_patch
