+++ upgrade_analysis/odoo_patch/addons/__init__.py
@@ -1,3 +1,3 @@
 from . import mrp
-from . import stock
 from . import point_of_sale
+from . import stock
