--- upgrade_analysis/odoo_patch/addons/point_of_sale/__init__.py
@@ -1,1 +1,0 @@
-from . import models
