--- upgrade_analysis/odoo_patch/addons/point_of_sale/models/__init__.py
@@ -1,1 +1,0 @@
-from . import pos_config
