+++ upgrade_analysis/odoo_patch/addons/stock/__init__.py
@@ -1,17 +1,10 @@
 from odoo.addons import stock
+from ...odoo_patch import OdooPatch
 
 
-def pre_init_hook(cr):
-    # <OpenUpgrade:REMOVE>
-    # don't uninstall data as this breaks the analysis
-    # Origin of this code is https://github.com/odoo/odoo/issues/22243
-    # env = api.Environment(cr, SUPERUSER_ID, {})
-    # env['ir.model.data'].search([
-    #     ('model', 'like', '%stock%'),
-    #     ('module', '=', 'stock')
-    # ]).unlink()
-    pass
-    # </OpenUpgrade>
+class PreInitHookPatch(OdooPatch):
+    target = stock
+    method_names = ['pre_init_hook']
 
-
-stock.pre_init_hook = pre_init_hook
+    def pre_init_hook(cr):
+        """ Don't unlink stock data on reinstall """
