+++ upgrade_analysis/odoo_patch/odoo/__init__.py
@@ -1,5 +1,3 @@
-from . import modules
-from . import service
-from . import tools
-from . import http
 from . import models
+from . import modules
+from . import tools
