+++ upgrade_analysis/odoo_patch/odoo/modules/__init__.py
@@ -1,12 +1,1 @@
-# Minor changes. (call to safe_eval changed)
-# otherwise : adapted to V14
-from . import graph
-
-# A lot of changes in the core functions.
-from . import loading
-
-# Adapted to V14
-from . import migration
-
-# Adapted to V14
 from . import registry
