+++ upgrade_analysis/odoo_patch/odoo/modules/registry.py
@@ -1,58 +1,29 @@
-# flake8: noqa
-# pylint: skip-file
+import logging
+from threading import current_thread
+from odoo import api, SUPERUSER_ID
+from ...odoo_patch import OdooPatch
+from .... import upgrade_log
+from odoo.modules.registry import Registry
 
-from collections import deque
-from contextlib import closing
-import odoo
-from odoo.tools.lru import LRU
-
-from odoo.modules import registry
+_logger = logging.getLogger(__name__)
 
 
-if True:
+class RegistryPatch(OdooPatch):
+    target = Registry
+    method_names = ['init_models']
 
-    def _init(self, db_name):
-        self.models = {}    # model name/model instance mapping
-        self._sql_constraints = set()
-        self._init = True
-        self._assertion_report = odoo.tests.runner.OdooTestResult()
-        self._fields_by_model = None
-        self._ordinary_tables = None
-        self._constraint_queue = deque()
-        self.__cache = LRU(8192)
+    def init_models(self, cr, model_names, context, install=True):
+        module_name = context['module']
+        _logger.debug('Logging models of module %s', module_name)
+        upg_registry = current_thread()._upgrade_registry
+        local_registry = {}
+        env = api.Environment(cr, SUPERUSER_ID, {})
+        for model in env.values():
+            if not model._auto:
+                continue
+            upgrade_log.log_model(model, local_registry)
+        upgrade_log.compare_registries(
+            cr, context['module'], upg_registry, local_registry)
 
-        # modules fully loaded (maintained during init phase by `loading` module)
-        self._init_modules = set()
-        self.updated_modules = []       # installed/updated modules
-        # <OpenUpgrade:ADD>
-        self.openupgrade_test_prefixes = {}
-        # </OpenUpgrade>
-        self.loaded_xmlids = set()
-
-        self.db_name = db_name
-        self._db = odoo.sql_db.db_connect(db_name)
-
-        # cursor for test mode; None means "normal" mode
-        self.test_cr = None
-        self.test_lock = None
-
-        # Indicates that the registry is
-        self.loaded = False             # whether all modules are loaded
-        self.ready = False              # whether everything is set up
-
-        # Inter-process signaling:
-        # The `base_registry_signaling` sequence indicates the whole registry
-        # must be reloaded.
-        # The `base_cache_signaling sequence` indicates all caches must be
-        # invalidated (i.e. cleared).
-        self.registry_sequence = None
-        self.cache_sequence = None
-
-        # Flags indicating invalidation of the registry or the cache.
-        self.registry_invalidated = False
-        self.cache_invalidated = False
-
-        with closing(self.cursor()) as cr:
-            self.has_unaccent = odoo.modules.db.has_unaccent(cr)
-
-registry.init = _init
+        return RegistryPatch.init_models._original_method(
+            self, cr, model_names, context, install=install)
