--- upgrade_analysis/odoo_patch/odoo/service/__init__.py
@@ -1,4 +1,0 @@
-# Import disabled, because the function run_unit_tests()
-# disappeared in V14.
-# TODO: OpenUpgrade Core maintainers : FIXME.
-# from . import server
