+++ upgrade_analysis/odoo_patch/odoo/tools/__init__.py
@@ -1,2 +1,1 @@
 from . import convert
-from . import view_validation
