+++ upgrade_analysis/views/view_upgrade_analysis.xml
@@ -37,6 +37,10 @@
                             attrs="{'readonly': [('state', '=', 'done')]}"
                         />
                         <field
+                            name="upgrade_path"
+                            attrs="{'readonly': [('state', '=', 'done')], 'invisible': [('write_files', '=', False)]}"
+                        />
+                        <field
                             name="analysis_date"
                             attrs="{'invisible': [('analysis_date', '=', False)]}"
                         />
